pyproject.toml
[project]
name = "oc_hooks"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
    "python-dotenv",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
